XILFPGA: MISRA-C:2012 R.10.1, 10.2, 10.3, 10.4 fix in xilfpga_library
This Patch modifies the source files in  xilfpga to be compliant with MISRA-C:2012 R.10.1,10.2,10.3,10.4

The files are:
xilfpga_pcap.h

The Rules R.10.1,10.2,10.3,10.4 says
 "Both operands of an operator in which usual arithmatic conversions
 are performed shall have the same essential type category"


Acked-for-series: Nava kishore Manne <nava.manne@xilinx.com>

+++ lib/sw_services/xilfpga/src/interface/zynqmp/xilfpga_pcap.h
@@ -242,8 +242,8 @@ extern "C" {
 #define XFPGA_PCAP_ERR_MASK			(0xFF00U)
 #define XFPGA_ERR_MODULE_MASK			(0xFFFF0000U)
 #define XFPGA_PCAP_UPDATE_ERR(XfpgaPcapErr, ModuleErr)		\
-		(((ModuleErr) << 16U) & XFPGA_ERR_MODULE_MASK) + \
-		(((XfpgaPcapErr) << 8U) & XFPGA_PCAP_ERR_MASK)
+		(((ModuleErr) << (u32)16U) & XFPGA_ERR_MODULE_MASK) + \
+		(((XfpgaPcapErr) << (u32)8U) & XFPGA_PCAP_ERR_MASK)
 
 #define XFPGA_STATE_MASK	0x00FF0000U
 #define XFPGA_STATE_SHIFT	16U
